# schema/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FAR-Lab API",
    "description": "FAR-Lab 对外 HTTP API（24§5）。V2 receipts 端点采用统一信封 { ok: true, data: T } + RFC 7807 错误响应（R-05）。",
    "version": "2026-06-27"
  },
...
  "paths": {
...
    "/api/v2/receipts": {
      "post": {
        "summary": "Create a receipt (idempotent by proofHash)",
        "tags": [
          "v2-receipts"
        ],
        "description": "Create a persisted receipt. Idempotent: same proofHash returns existing receiptId with 200 + idempotent=true. New receipt returns 201 + idempotent=false. Response: { ok: true, data: { receiptId, idempotent } }.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "proofHash": {
                    "type": "string",
                    "minLength": 1
                  },
                  "schemaVersion": {
                    "type": "string",
                    "minLength": 1
                  },
                  "claimId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "claimText": {
                    "type": "string",
                    "minLength": 1
                  },
                  "verdict": {
                    "type": "string",
                    "minLength": 1
                  },
                  "manifestMembers": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "kind": {
                          "type": "string",
                          "minLength": 1
                        },
                        "digest": {
                          "type": "string",
                          "minLength": 1
                        },
                        "sizeBytes": {
                          "type": "number",
                          "minimum": 0
                        }
                      },
                      "required": [
                        "kind",
                        "digest",
                        "sizeBytes"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "contractBindings": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "bindingSetJson": {
                          "type": "string",
                          "minLength": 1
                        },
                        "digest": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": [
                        "bindingSetJson",
                        "digest"
                      ],
                      "additionalProperties": false
                    }
                  }
                },
                "required": [
                  "proofHash",
                  "schemaVersion",
                  "claimId",
                  "claimText",
                  "verdict"
                ],
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "receiptId": {
                          "type": "string"
                        },
                        "idempotent": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "receiptId",
                        "idempotent"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "ok",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "receiptId": {
                          "type": "string"
                        },
                        "idempotent": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "receiptId",
                        "idempotent"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "ok",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "source_anchor": {
                      "type": "object",
                      "properties": {
                        "fileId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "stageId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "callRecordId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        }
                      },
                      "required": [
                        "fileId",
                        "stageId",
                        "callRecordId"
                      ],
                      "additionalProperties": false
                    },
                    "detail": {}
                  },
                  "required": [
                    "error_code",
                    "message",
                    "source_anchor"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List receipts (limit/offset pagination, optional claimId filter)",
        "tags": [
          "v2-receipts"
        ],
        "description": "List persisted receipts with limit/offset pagination. Response: { ok: true, data: { receipts, total, limit, offset } }. limit: 1..100 (default 20), offset: >=0 (default 0), claimId: optional exact-match filter (runId of a wizard run).",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "in": "query",
            "name": "offset",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            },
            "in": "query",
            "name": "claimId",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "receipts": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "claimId": {
                                "type": "string"
                              },
                              "claimText": {
                                "type": "string"
                              },
                              "verdict": {
                                "type": "string"
                              },
                              "proofHash": {
                                "type": "string"
                              },
                              "schemaVersion": {
                                "type": "string"
                              },
                              "createdAt": {
                                "type": "string"
                              },
                              "receiptStanding": {
                                "type": "string"
                              },
                              "preservationStatus": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "claimId",
                              "claimText",
                              "verdict",
                              "proofHash",
                              "schemaVersion",
                              "createdAt",
                              "receiptStanding",
                              "preservationStatus"
                            ],
                            "additionalProperties": true
                          }
                        },
                        "total": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "offset": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "receipts",
                        "total",
                        "limit",
                        "offset"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "ok",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "source_anchor": {
                      "type": "object",
                      "properties": {
                        "fileId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "stageId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "callRecordId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        }
                      },
                      "required": [
                        "fileId",
                        "stageId",
                        "callRecordId"
                      ],
                      "additionalProperties": false
                    },
                    "detail": {}
                  },
                  "required": [
                    "error_code",
                    "message",
                    "source_anchor"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/receipts/{id}/verify": {
      "get": {
        "summary": "Re-verify a receipt (run + persist six-dimension result)",
        "tags": [
          "v2-receipts"
        ],
        "description": "Run V2 six-dimension verification on a persisted receipt, persist the result, return it. Response: { ok: true, data: { verification, display, allPass } }. 404 if not found (RFC 7807).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "verification": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "display": {
                          "type": "string"
                        },
                        "allPass": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "verification",
                        "display",
                        "allPass"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "ok",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "source_anchor": {
                      "type": "object",
                      "properties": {
                        "fileId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "stageId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "callRecordId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        }
                      },
                      "required": [
                        "fileId",
                        "stageId",
                        "callRecordId"
                      ],
                      "additionalProperties": false
                    },
                    "detail": {}
                  },
                  "required": [
                    "error_code",
                    "message",
                    "source_anchor"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    }
  },
...
}